Rehome sync helpers per review: BaaS query to BaasPublishOpsMixin, supersede to its caller
- get_baas_publish_progress moves from ProgressSyncMixin to
  BaasPublishOpsMixin — it is a BaaS-layer publish-workflow query, the same
  concern as approve_baas_publish.
- The sync entry points read publish records through the existing public
  get_publish_record accessor instead of raw _publish_service.get_publish_by_id.
- Drop PublishExtMixin._record_sync_success: it was only a retry-flag pop plus
  one _update_publish_status call, so _handle_sync_success now does those two
  steps itself (persistence still goes through the ext_state plumbing).
- _mark_previous_publish_superseded moves to ProgressSyncMixin — it is part of
  the sync-success handling and that is its only caller; a dedicated mixin for
  one method would be noise. Its write stays on the _update_publish_status
  plumbing, so the single-writer invariant holds.
