+++ src/backend/src/agentclaw/community/core/service_bot/services/publish_flow/baas_publish_ops_mixin.py
@@ -1,5 +1,7 @@
-"""Shared BaaS publish approval, mixed in."""
+"""Shared BaaS-layer publish ops, mixed in."""
 from __future__ import annotations
+
+from typing import Any, Dict
 
 from agentclaw.community.core.service_bot.types import PublishStage
 from agentclaw.community.log import get_logger
@@ -8,12 +10,12 @@ logger = get_logger()
 
 
 class BaasPublishOpsMixin:
-    """Shared BaaS publish approval, mixed in.
+    """Shared BaaS-layer publish ops, mixed in.
 
-    This mixin owns only the BaaS-layer approve call. Creating the device binding
-    and writing the publish record's ext/status are separate concerns that live on
-    the facade (``create_release_binding`` / ``record_release_ext``); the release
-    runner invokes the three steps in sequence.
+    Owns the calls against the BaaS publish workflow — approve and progress
+    query. Creating the device binding and writing the publish record's
+    ext/status are separate concerns that live on ``DeviceBindingMixin`` /
+    ``PublishExtMixin``; the release runner invokes the steps in sequence.
     """
 
     def approve_baas_publish(
@@ -56,3 +58,27 @@ class BaasPublishOpsMixin:
                 f"stage={stage.value}, error={e}, continuing..."
             )
             return False
+
+    def get_baas_publish_progress(
+        self,
+        *,
+        baas_publish_id: int,
+        include_devices: bool = True,
+    ) -> Dict[str, Any]:
+        """Query BaaS publish progress."""
+        logger.info(
+            f"[PublishFlowService.get_baas_publish_progress] Query BaaS progress: "
+            f"baas_publish_id={baas_publish_id}, include_devices={include_devices}"
+        )
+        try:
+            return self._baas_service.get_publish_progress(
+                publish_id=int(baas_publish_id),
+                include_devices=include_devices,
+            )
+        except Exception as e:
+            logger.error(
+                f"[PublishFlowService.get_baas_publish_progress] "
+                f"Failed to get BaaS publish progress: baas_publish_id={baas_publish_id}, "
+                f"error={e}"
+            )
+            raise
